fix(build): fix generated package.json to not include Custom UI deps
replace package.json generator with custom implementation that makes sure that Custom UI dependencies are not included in the generated Forge app package.json

Closes: #13
